Drop `issue()` from `Issuer` interface
